+ Enum State aggiunto + Iniziata parte su inserimento di partecipanti

+++ ManageExpenses/app/src/main/java/com/example/gianluca/manageexpenses/ExpenseOperations.java
@@ -7,11 +7,14 @@ package com.example.gianluca.manageexpenses;
 public class ExpenseOperations {
 
 
-    public static double divide(int dividendo,int divisore){
+
+
+    public static double divide(double dividendo,double divisore){
 
         return dividendo/divisore;
 
     }
+
 
 
 }
